Plugins apache_*: tolerate URL without placeholder for port
The plugin documentation requires the placeholder "%d" to be present in
the configured URL.

Sadly this can lead to confusing warning messages:
  Redundant argument in sprintf

From now on it is allowed to provide a URL without a port placeholder.
This prevents the above warning messages.
Obviously users with multiple ports will understand, that they need to
add a port placeholder in order to access multiple instances.

Closes: #1319