munin-node.postinst: remove unused conditional
Previously the existence (even if empty) of the stdout-redirected file
was tested in "configure".  This test always succeeded - even in case of
a failing process execution, since the file is created by the shell, not
by the process.

Additionally the branches of the conditional were problematic, since in
case of a assumed failure, potentially interesting error messages would
have been hidden.