Move "munin-node-configure" test from Makefile to travis checks
Commit a77b86f4fc97d3539 added a for the "autogain" feature of plugins.
Sadly this test relies on the local test environment managed via the
scripts below "dev_scripts/".  These scripts require a "make clean" and
therefore reset the current build state.  This side-effect is not
feasible in combination with an external build environment (e.g. for
building Debian packages).

Thus the "autoconf" test is now only executed in the context of the CI
test procedure.  Side-effects are acceptable in this throw-away
environment.