Plugins: force execution in a shell for "command -v"
The behaviour of perl's backticks (or "system" function) depends on the
content of the string: it is executed via "execvp", if no shell meta
characters are found. Otherwise it is executed by the shell.
In order to use the portable "command -v", we need to enforce the
interpretation of the command via the shell.