Make sensors_ temperature plugin work on musl libc
musl libc (e.g. Alpine Linux, OpenWrt) behaves slightly differently than
glibc when LC_ALL=C is passed to the sensors command. Instead of
substituting the degree symbol for a space, it replaces it with an
asterisk. Adjust the regexes accordingly.

Also, change the autoconf regex to be consistent with the ones used to
actually parse the 'sensors' output.