SpoolReader: process spooled content lines separately
Previously the full content of the spool data was handled in memory,
before it was printed at once.
Now every line is printed directly, when it is processed.
Thus the munin-async process does not consume a relevant amount of
memory anymore.

The issue was reported and tested by Sephtex via IRC, thanks!