CFED handling: fix translation of fieldnames to RRD field names
The helper function "orig_to_cdef" is used in contexts, where its result
is used as a reference to an RRD field name.  Thus the result of the
function needs to be converted into an RRD field name (via
"get_field_name").
This is relevant only for long field names (> 19 characters).

See #1377