plugin.sh: remove conversation log regarding shell redirects
The comment was added in c9acbf34953 and is probably supposed to be
helpful for the future implementation of a new function in plugin.sh.
But the comment does not seem to be related to the simultaneous changes
in the perl module.  It is probably of no use anymore.

Thanks, Sander Bos!