apt_all: remove reference to "update" argument for cron job
Before commit 904a84e392ed2 the plugin allowed to be called with an
"update" argument in order to trigger a package cache update.

The plugin does not support this anymore.
Instead the infrastructure provided by the system ("APT::Periodic")
should be used instead.
This is now clarified in the plugin documentation.