ntp_: do not mask required variable from Net::IP
Using "my" will declare a new variable at (perl) compile time, effectively
masking $IP_IDENTICAL from Net::IP which is used later. When executed
on the shell, that will look like this:

```
Use of uninitialized value $IP_IDENTICAL in numeric eq (==) at ./ntp_XXX line 175.
delay.value 0.000
offset.value +0.000
jitter.value 0.000
```

Instead we need to make sure to reference $Net::IP::IP_IDENTICAL if
available (if it is not available, it's no problem: plugin net_ will
refuse to autoconfigure and fail hard anyways if configured manually).
The easy way for this would be to use "our", but that is rather inelegant
and sets a bad example (the scoping with "our" is more complex than with
"my").
This patch copies $Net::IP::IP_IDENTICAL to a true local variable if
available - if not, the eval block bombs out anyways. As $IP_IDENTICAL
is used exactly once, that would result in a warning, which is surpressed
in that block only.