tests: Avoid remaining race with cron shutdown
commit 5379a779 ("tests: completely disable all cron services and suppress
errors") attempted to avoid a race between testing for generated files
and the cron-triggered script that generates them. But I noticed this test
still seems to fail some times in Ubuntu's CI. I tried running it in a
loop locally and see a failure rate of ~4% (20/531).

My theory is that there's still a race window between when cron has
started munin-cron, and munin-cron starting munin-update. So instead
of just killing off munin-update, let's target the root of the problem -
munin-cron - and any other processes (such as munin-update) that it may
have started in its process group.

So far, I've 188 runs w/ 0 failures, so I'm hopeful this has finally
closed the race.

Signed-off-by: dann frazier <dannf@debian.org>
Signed-off-by: Holger Levsen <holger@layer-acht.org>