master: get graph width/height for html in graph-cgi mode also
When generating html, if graph mode is cron, munin looks for
the generated graph png files (from the previous run) to get
width and height for the img elements, so that the page doesn't
jump around for the few seconds it takes to download all the graph
images. This can be done if graph mode is cgi, also, by looking
in the cgitmpdir. Sometimes graph images won't be found, but if
they are found they can be used.

-----------------

This is admittedly a "hacky" implementation, and targetting
the last stable branch instead of master. It is offered as an
illustration of an idea that would be nice to have, if implemented
properly. (The hacky implementation does work :)