Sed in the blib directory before install
Sed when not making packages resulted in `grep -rl --null "@@" .`
which includes _everything_ also the .git directory and one of the git
.pack files contained a match for the sed expression thereby corrupting
the git checkout.  And also not making any substitutions in the
installed files in /usr/local...  So sed has to be done prior to install
in the blib directory.