Allow using unix sockets in the form unix:///run/muninnode.sock as address
On a normal host with lots of docker containers, they can easily connect to the host's port 4949 and reveal a lof of information about internal structures. This patch allows munin to
connect to unix sockets instead which are much easier to protect. Munin mode so far doesn't support unix sockets, but muninlite can be configured (using inetd/xinetd) to do this.