Prevent log spam with nut plugins
On Debian / Ubuntu, nut is linked against libnss3 which causes a warning
message to be emitted when upsc is run if SSL isn't being used. This
leads to errors during every poll like:

2024/09/04-21:05:07 [299208] Error output from nut_misc:
2024/09/04-21:05:07 [299208]         Init SSL without certificate database

nut 2.8.1, which has started making its way to distros, adds a new
NUT_QUIET_INIT_SSL environment variable (specifically designed for uses
such as this) which allows silencing this warning.

Add it to the nut plugins. It's a no-op for older nut versions.