Enable wget --no-check-certificate option in http_loadtime plugin
This patch disables SSL/TLS certificate validation in the http_loadtime plugin
when wget establishes an HTTPS connection. This prevents the plugin from
considering an untrusted certificate a failure. Which allows the plugin to work
with HTTPS protocol servers, even when they supply a self-signed, expired, or
otherwise invalid certificate.